pyproject.toml
[tool.poetry]
name = "chrome-remote"
version = "0.0.3"
description = ""
authors = ["dadevel <dadevel@disroot.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
chrome-remote = "chrome_remote.main:main"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
websocket-client = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
